+++ agent/bindings/kotlin/lib/src/main/kotlin/com/sourcegraph/cody/protocol_generated/GetDocumentsParams.kt
@@ -1,8 +1,7 @@
 @file:Suppress("FunctionName", "ClassName", "unused", "EnumEntryName", "UnusedImport")
 package com.sourcegraph.cody.protocol_generated;
 
-data class SearchPanelSnippet(
-  val contents: String,
-  val range: RangeData,
+data class GetDocumentsParams(
+  val uris: List<String>? = null,
 )