+++ vscode/src/jsonrpc/agent-protocol.ts
@@ -2,6 +2,7 @@ import type {
     AuthStatus,
     BillingCategory,
     BillingProduct,
+    CodyCommand,
     ContextFilters,
     CurrentUserCodySubscription,
     Model,
@@ -84,6 +85,9 @@ export type ClientRequests = {
 
     // Trigger custom commands that could be a chat-based command or an edit command.
     'commands/custom': [{ key: string }, CustomCommandResult]
+
+    // A list of available custom commands stored in .cody/commands.json.
+    'customCommands/list': [null, CodyCommand[]]
 
     // Trigger commands that edit the code.
     'editCommands/code': [
@@ -180,6 +184,8 @@ export type ClientRequests = {
     'testing/closestPostData': [{ url: string; postData: string }, { closestBody: string }]
     'testing/memoryUsage': [null, { usage: MemoryUsage }]
     'testing/awaitPendingPromises': [null, null]
+    // Retrieve the Agent's copy of workspace documents, for testing/validation.
+    'testing/workspaceDocuments': [GetDocumentsParams, GetDocumentsResult]
     // Returns diagnostics for the given URI. Lives under `testing/` instead of
     // standalone `diagnostics/` because it only works for TypeScript files.
     'testing/diagnostics': [{ uri: string }, { diagnostics: ProtocolDiagnostic[] }]
@@ -200,6 +206,8 @@ export type ClientRequests = {
 
     // Returns the current authentication status without making changes to it.
     'extensionConfiguration/status': [null, AuthStatus | null]
+
+    'textDocument/change': [ProtocolTextDocument, { success: boolean }]
 
     // Run attribution search for a code snippet displayed in chat.
     // Attribution is an enterprise feature which allows to look for code generated
@@ -922,4 +930,16 @@ export interface ProtocolCodeAction {
           }
         | undefined
         | null
+}
+
+/**
+ * Omitting uris parameter will retrieve all open documents for the
+ * current workspace root.
+ */
+export interface GetDocumentsParams {
+    uris?: string[] | undefined | null
+}
+
+export interface GetDocumentsResult {
+    documents: ProtocolTextDocument[]
 }